refactor(layout): inline AppShell into app layout, use Opal sidebar width tokens
Delete AppShell.tsx and move its sidebar fold state directly into
app/app/layout.tsx. Sidebars now reference --sidebar-width-expanded and
--sidebar-width-folded from Opal instead of hardcoded px values. Body
background set via @apply bg-background to match Onyx.
