+++ frontend/src/app/globals.css
@@ -12,6 +12,12 @@ body {
     scrollbar-gutter: stable;
 }
 
+@layer base {
+  body {
+    @apply bg-background;
+  }
+}
+
 /* Tell the browser which color scheme is active so native UI (scrollbars,
    autofill, date pickers) uses matching light/dark variants. */
 :root {
@@ -56,17 +62,17 @@ body {
 }
 /* Disabled primary action — neutral grey, never Opal's default blue tint. */
 .interactive[data-interactive-variant="action"][data-interactive-prominence="primary"][data-disabled] {
-  background-color: var(--background-tint-04) !important;
-  --interactive-foreground: var(--text-03);
-  --interactive-foreground-icon: var(--text-03);
+    background-color: var(--background-tint-04) !important;
+    --interactive-foreground: var(--text-03);
+    --interactive-foreground-icon: var(--text-03);
 }
 
 /* Selected menu/select items use the near-black accent text, not Opal's blue. */
 .interactive[data-interactive-variant="select-light"][data-interactive-state="selected"],
 .interactive[data-interactive-variant="select-heavy"][data-interactive-state="selected"],
 .interactive[data-interactive-variant="select-tinted"][data-interactive-state="selected"] {
-  --interactive-foreground: var(--text-05) !important;
-  --interactive-foreground-icon: var(--text-05) !important;
+    --interactive-foreground: var(--text-05) !important;
+    --interactive-foreground-icon: var(--text-05) !important;
 }
 
 /* ── Horizontal scroll containers ─────────────────────────────────────────
@@ -102,101 +108,101 @@ body {
    lists, code, and tables render as flat unstyled text.
    ───────────────────────────────────────────────────────────────────── */
 .markdown {
-  color: var(--text-05);
-  line-height: 1.6;
+    color: var(--text-05);
+    line-height: 1.6;
 }
 .markdown h1,
 .markdown h2,
 .markdown h3,
 .markdown h4 {
-  color: var(--text-05);
-  font-weight: 600;
-  line-height: 1.3;
+    color: var(--text-05);
+    font-weight: 600;
+    line-height: 1.3;
 }
 .markdown h1 {
-  font-size: 1.875em;
-  margin: 28px 0 10px;
+    font-size: 1.875em;
+    margin: 28px 0 10px;
 }
 .markdown h2 {
-  font-size: 1.5em;
-  margin: 24px 0 8px;
+    font-size: 1.5em;
+    margin: 24px 0 8px;
 }
 .markdown h3 {
-  font-size: 1.25em;
-  margin: 20px 0 6px;
+    font-size: 1.25em;
+    margin: 20px 0 6px;
 }
 .markdown h4 {
-  font-size: 1.1em;
-  margin: 16px 0 4px;
+    font-size: 1.1em;
+    margin: 16px 0 4px;
 }
 .markdown p {
-  margin: 10px 0;
+    margin: 10px 0;
 }
 .markdown ul,
 .markdown ol {
-  margin: 10px 0;
-  padding-left: 24px;
+    margin: 10px 0;
+    padding-left: 24px;
 }
 .markdown ul {
-  list-style: disc;
+    list-style: disc;
 }
 .markdown ol {
-  list-style: decimal;
+    list-style: decimal;
 }
 .markdown li {
-  margin: 4px 0;
+    margin: 4px 0;
 }
 .markdown a {
-  color: var(--text-05);
-  text-decoration: underline;
+    color: var(--text-05);
+    text-decoration: underline;
 }
 .markdown a:hover {
-  color: var(--text-04);
+    color: var(--text-04);
 }
 .markdown code {
-  background: var(--background-tint-02);
-  color: var(--text-05);
-  border: 1px solid var(--border-01);
-  border-radius: var(--border-radius-04);
-  padding: 1px 6px;
-  font-size: 0.92em;
+    background: var(--background-tint-02);
+    color: var(--text-05);
+    border: 1px solid var(--border-01);
+    border-radius: var(--border-radius-04);
+    padding: 1px 6px;
+    font-size: 0.92em;
 }
 .markdown pre {
-  background: var(--background-tint-02);
-  color: var(--text-05);
-  border: 1px solid var(--border-01);
-  border-radius: var(--border-radius-04);
-  padding: 12px 14px;
-  overflow-x: auto;
+    background: var(--background-tint-02);
+    color: var(--text-05);
+    border: 1px solid var(--border-01);
+    border-radius: var(--border-radius-04);
+    padding: 12px 14px;
+    overflow-x: auto;
 }
 .markdown pre code {
-  background: transparent;
-  border: none;
-  padding: 0;
-  font-size: inherit;
+    background: transparent;
+    border: none;
+    padding: 0;
+    font-size: inherit;
 }
 .markdown blockquote {
-  border-left: 3px solid var(--border-02);
-  margin: 12px 0;
-  padding: 4px 12px;
-  color: var(--text-04);
-  background: var(--background-tint-01);
+    border-left: 3px solid var(--border-02);
+    margin: 12px 0;
+    padding: 4px 12px;
+    color: var(--text-04);
+    background: var(--background-tint-01);
 }
 .markdown table {
-  border-collapse: collapse;
+    border-collapse: collapse;
 }
 .markdown th,
 .markdown td {
-  border: 1px solid var(--border-01);
-  padding: 6px 10px;
+    border: 1px solid var(--border-01);
+    padding: 6px 10px;
 }
 .markdown th {
-  background: var(--background-tint-02);
+    background: var(--background-tint-02);
 }
 .markdown hr {
-  border: none;
-  border-top: 1px solid var(--border-01);
-  margin: 16px 0;
+    border: none;
+    border-top: 1px solid var(--border-01);
+    margin: 16px 0;
 }
 
 /* ── Inline comment highlights (CSS Custom Highlight API) ─────────────────
@@ -204,8 +210,24 @@ body {
    Yellow/orange are illustrative decoration — raw hex is intentional here.
    ───────────────────────────────────────────────────────────────────── */
 ::highlight(wiki-comment) {
-  background-color: rgba(255, 212, 0, 0.38);
+    background-color: rgba(255, 212, 0, 0.38);
 }
 ::highlight(wiki-comment-active) {
-  background-color: rgba(255, 159, 28, 0.6);
+    background-color: rgba(255, 159, 28, 0.6);
+}
+
+@utility dbg-red {
+    /* DEBUGGING UTILITIES
+  If you ever want to highlight a component for debugging purposes, just type in `className="dbg-red ..."`, and a red box should appear around it.
+  This helps with placing things properly on the screen and seeing how they look during active development.
+  */
+    border: 1px solid red;
+}
+
+@utility dbg-blue {
+    border: 1px solid blue;
+}
+
+@utility dbg-green {
+    border: 1px solid green;
 }
